+++ make/common/db/registry-flag.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+touch /var/lib/mysql/created_in_mariadb.flag
+echo "dumped flag for MariaDB"