+++ tools/migration/changelog.md
@@ -54,9 +54,10 @@ Changelog for harbor database schema
 
   - create table `project_metadata`
   - insert data into table `project_metadata`
+  - delete column `public` from table `project`
+  - add column `insecure` to table `replication_target`
 
 ## 1.4.0
   - create table `replication_filter`
   - create table `replication_filter_type`
   - insert data into table `replication_filter_type`
-